Cu crear curso

+++ src/controllers/course.controller.ts
@@ -6,10 +6,8 @@ import {
   validateCourseToPatch,
   validateSearchByTitle,
 } from "./../schemas/course.schema.js";
-import { Topic } from "../entities/topic.entity.js";
-import { Level } from "../entities/level.entity.js";
+import { ZodError } from "zod";
 import { CoursePurchaseRecord } from "../entities/coursePurchaseRecord.entity.js";
-
 const em = orm.em;
 em.getRepository(Course);
 function sanitizeCourseInput(req: Request, res: Response, next: NextFunction) {
@@ -78,9 +76,17 @@ async function add(req: Request, res: Response) {
     const validCourse = validateCourse(req.body.sanitizedInput);
     const course = em.create(Course, { ...validCourse, createdAt: new Date() });
     await em.flush();
-    res.status(201).json({ message: "Course created", data: course });
+    const courseCreated = em.getReference(Course, course.id);
+    res
+      .status(201)
+      .json({ message: "Course created", data: { courseCreated } });
   } catch (error: any) {
-    res.status(500).json({ message: error.message });
+    if (error instanceof ZodError) {
+      return res
+        .status(400)
+        .json(error.issues.map((issue) => ({ message: issue.message })));
+    }
+    res.status(500).send({ message: error.message });
   }
 }
 
@@ -88,12 +94,10 @@ async function update(req: Request, res: Response) {
   try {
     const id = Number.parseInt(req.params.id);
     const course = em.getReference(Course, id);
-    let courseUpdated;
-    if (req.method === "PATCH") {
-      courseUpdated = validateCourseToPatch(req.body.sanitizedInput);
-    } else {
-      courseUpdated = validateCourse(req.body.sanitizedInput);
-    }
+    const courseUpdated =
+      req.method === "PATCH"
+        ? validateCourseToPatch(req.body.sanitizedInput)
+        : validateCourse(req.body.sanitizedInput);
     em.assign(course, courseUpdated);
     await em.flush();
     res.status(200).json({ message: "Course updated", data: course });
@@ -106,8 +110,17 @@ async function remove(req: Request, res: Response) {
   try {
     const id = Number.parseInt(req.params.id);
     const course = em.getReference(Course, id);
-    await em.removeAndFlush(course);
-    res.status(204).json({ message: "Course deleted" });
+    const purchaseRecordCount = await em.count(CoursePurchaseRecord, {
+      course,
+    });
+    if (purchaseRecordCount > 0) {
+      course.isActive = false;
+      await em.flush();
+      return res.status(200).json({ message: "Course deactivated" });
+    } else {
+      await em.removeAndFlush(course);
+      res.status(204).json({ message: "Course deleted" });
+    }
   } catch (error: any) {
     res.status(500).json({ message: error.message });
   }
